Resource Copy: Cache created resources & views
When we copy a resource we create a new resource to hold the copy, and
we may also need to create a new view in some cases. Creating these
every time is wasteful, and can cause a very noticeable performance
bottleneck if hunting is turned on as the stat collection code sees
these as new resources to hash.

To optimise this we hold on to any resources and views we create and
will try to reuse them when possible, only recreating them if they are
not compatible with the resource being copied. For example, if we are
copying a depth buffer we would usually expect to be able to reuse the
same resource unless the swap chain was resized for some reason, which
should be rare in most games.

There may still be pathological cases. For example, if we were copying a
texture from a UI shader that is used for multiple UI elements with
different sized textures we would throw away the cached resource every
time. Another example may be a game that uses the same shader with
multiple render targets of different sizes, such as for an avatar, or a
game using multiple views like the cameraphone in DreadOut. If these
become a problem in practice we can look into making a more
sophisticated cache.