Resource Copy: Implement 'unless_null' option & change default
If we are asked to copy a resource that doesn't exist, set the
destination to NULL by default - that way, if someone makes a mistake in
the d3dx.ini it will be more obvious than the current behaviour where an
existing resource would remain assigned.

Add an 'unless_null' option to return to the previous behaviour if it is
required for some reason.