Resource Copy: Add custom resource sections
These can be used as an intermediate storage when copying a resource,
like a temporary variable. This allows resources to be copied between
unrelated draw calls explicitly, which among other things means we can
now accurately fix physical lighting effects in Unity 5 games.

It also gives us a starting point which we could use to create resources
entirely from scratch if we needed to, say to inject a texture loaded
from a file, or to add an extra render target to store some calculated
value, but currently it's just for intermediate storage.