Resource copying: Hook up UAVs (minimally tested)
This will allow UAVs to be copied around. It's only been minimally
tested as there aren't many games that use UAVs that we can prove it out
with. I see this as likely being more useful if we start injecting our
own shaders, and we can debug it as necessary then.

I'm not entirely certain if I've understood the documentation for the
pixel shader case - it's not clear if the start UAV slot must be equal
to the number of RTVs, or can be greater than that. It's also not clear
if setting one UAV will unbind others. I'm not overly worried about that
though - setting a UAV is something we are unlikely to be doing unless
we have injected our own shader to go with it.

Currently there is no way to set the hidden counter for append/consume
and RWStructuredBuffers structured buffers. This probably doesn't matter
yet, but will be important if we ever create custom buffers of these
types, which I do have some plans for. This is a little tricky as it can
only be set at bind time, but that is not the most suitable time to
initialise it.