Frame Analysis: Add options to dump vertex and index buffers
It might be handy to get an idea of what is in the vertex buffers to
determine what data is passed to the vertex shaders by the game. This
could be useful if we pass a vertex buffer to a shader in a constant
buffer to enable looking up information from other vertices.

This currently does not dump out the input layout, which would be
useful to help interpret the vertex buffers. The input layout is
separate though, and doesn't provide an interface to query anything of
interest so we would have to wrap it to get this.

It also doesn't dump out some of the extra information we get along with
the buffers - strides, offsets, and format.