Allow duplicate keys in ini sections implementing a command list
Since the command lists process every line in these sections, it it
legal for lines to be duplicated, so don't warn about duplicate keys in
these sections. In particular, we are likely to want multiple
checktextureoverride lines in a ShaderOverride section, but there are
other possible scenarios where we might have a duplicate line.

But - we still want to detect and warn about duplicate keys that are not
part of the command list, so add some duplicate detection into the
ParseCommandList() to get the best of both worlds.

Note that this necessitates listing all command list sections - since
this list will inevitably not be kept up to date leading to us warning
about duplicates we shouldn't, add a check in ParseCommandList to ensure
it is only called for sections in that list, and terminate the program
if we forget... that should prevent a few bugs ;-)