Resource Copying: Skip cache when copying to a CustomResource
When we are copying a resource to a CustomResource we can use the
Resource and View in the CustomResource directly instead yet another
intermediate resource.

This means that if we are copying something to a single custom resource
from a bunch of shaders we will only end up with one copy in the custom
resource, whereas before we would get an extra copy in the cache of each
copy operation structure.