Add a [Present] section with a command list that runs off Present()
There might be some things we want to do at the start/end of each frame.
For example, we might want to clear a custom resource so that it is only
valid if it gets set during a frame, or we might want to set an IniParam
that we are using for scene detection to 0.

Maybe later when we have shader injection working we might want to
inject a custom shader at this point to visualise something...

Add a [Present] section to the d3dx.ini which holds a command list
to give us the greatest flexibility to do what we like.