Resource copying: Don't throw away CustomResource when set to null
If we set a custom resource to null, we will always create a new one the
next time it is set to something valid. This is wasteful, and can cause
a bottleneck in the stat collection, so is best avoided.

Add a flag to the CustomResources to indicate when they are set to null,
but keep the resource & view around so we don't need to recreate them
later.