Shader copy: Implement [Resource] max_copies_per_frame
This can be used as a simple way to reduce the number of copy operations
we do in a frame if we obtain a resource from multiple shaders/draw
calls.

A global frame counter is added, which we can use to check if a resource
has been copied so far this frame.

Note that a resource can still be set to null during a frame, either
explicitly or by copying a null resource. This is intentional as we
might be explicitly clearing a resource from [Present]. If it is not
desired to set it to null by copying a null resource, the "unless_null"
keyword can be used in the copy - this may be necessary since if we have
hit the maximum number of copies for a frame we won't set it again until
the next frame.