Restore NeedOriginalShader condition on G->hunting
After we allowed hunting to be soft enabled we removed it as one of the
conditions to keeping a copy of the original shader around for
marking_mode=original or show_original.

We have since added a distinction between hunting being soft or hard
disabled, so add the condition back to avoid saving original shaders if
hunting is hard disabled.

Keep the original shaders only if we need it for depth buffer or partner
shader filtering, or if hunting either enabled / soft disabled and one
of the two reasons we may need the original is true, or the config file
is reloadable.