Fix marking a shader interfering with marking_mode=original or show_original
With marking_mode=original, when a shader was marked 3DMigoto would stop
showing the original shader, and show_original would also stop working
for that shader.

The reason was that once the shader was in the mReloadedShaders map it
then pShader would be substituted immediately, and therefore would no
longer be able to find the shader in the originalShaders map since the
key used there was the original shader and not the replaced one.

Fix it using a second variable to distinguish the original shader passed
from the game with the one we are replacing it with, and always look up
maps using the original one.