Add a QueryInterface call to HackerDXGIAdapter to return HackerDXGIAdapter1.
COM uses this bizarre type-coercion/upcasting mechanism through QueryInterface, and the game DyingLight is using a call to DXGIAdapter::QueryInterface(DXGIAdapter1) to get the Adapter1.

This change now returns a wrapped HackerDXGIAdapter1 for that case, and also returns an error for the HackerDXGIAdapter2 case (as being on that requires the evil update, following our model.)

This fixes the problem in DyingLight, game loads now, and overlay is working in 1.2.26 version.