Fix GetImmediateContext() returning an invalid pointer
In Far Cry 4 a several devices are created, some of which are created
along with an immediate context, but others are not.

Those that do not have an immediate context already created will create
one in the GetImmediateContext() call, which we wrap. If that context's
reference count drops to 0, we attempt to set the mHackerContext member
of the device to NULL, but due to a bug we do not have the pointer to
the device set in that case and consequently leave it as an invalid
pointer.

The next time GetImmediateContext() is called, it will return the
invalid pointer, which could lead to a double free or other crash:

->D3D11CreateDevice returned device handle = 0000000018E9EE18, context handle = 0000000000000000
  HackerDevice 0000000015677AB0 created to wrap 0000000018E9EE18

...

HackerDevice::GetImmediateContext(class HackerDevice@0000000015677AB0) called with:00000000160B83A0
*** HackerContext missing at HackerDevice::GetImmediateContext
  HackerContext 0000000018E2EC40 created to wrap 0000000018E9F870 (0000000018E9F870)
  returns handle = 0000000018E2EC40

HackerContext::Release counter=0, this=0000000018E2EC40
  deleting self

HackerDevice::GetImmediateContext(class HackerDevice@0000000015677AB0) called with:00000000160B8328
  returns handle = 0000000018E2EC40

This occurs several times while launching Far Cry 4 - during the spash
screen and at least once more before the menu has loaded, and sometimes
this causes it to crash on launch.

This patch calls SetHackerDevice() on the newly wrapped context in
GetImmediateContext() so that it will be able to NULL out the context
pointer in the device when it is freed.

It also makes sure that the context will only set the device's
mOrigContext to NULL if it matches the context being released, which
should prevent deferred contexts from inadvertently clearing it.

It also adds some extra debug output to hopefully catch similar issues
occurring in the future - e.g. it might be possible in some
multi-threaded scenarios for this to still occur due to a race (this
turned out not to be the case in FC4 as separate threads were accessing
separate devices), and if we ever support Device1/Context1 in the future
it is likely a similar issue would exist there.