Fix bugs when stereo is disabled in the control panel
- Creating the ini params resource is no longer conditional on stereo
- Creating the pinking shader is no longer conditional on stereo
- Crash on config reload with stereo disabled is fixed

All of these were caused by the same issue - if stereo was disabled in
the control panel, or we were not running on an nvidia card we would
skip too much of the init sequence and would never create the ini params
resources or pinking shader, even though neither of these are dependent
on stereo.

The stereo params resource is currently still dependent on stereo being
enabled. This should generally be ok as reading from an unassigned
resource should return 0 and that should result in any corrections
based on separation*something becoming 0, but it's worth considering if
we want to create a stereo params resource anyway since it also contains
tuning values and the screen resolution. I think the former has been
broken for some time, and the later is available in the ini params
texture as well (albiet not exactly the same), so this did not seem
worthwhile at this stage.