Remove double logging of HackerDevice::Release
Hitting a crash in here in Far Cry 4 some of the time in windowed mode
(sometimes after 10 minutes, sometimes while loading the menu). Thought
it was a double free based on the logs, but it turns out the Release log
message was just being printed twice. Still not sure of the actual cause
of the crash, but let's at least clear up the confusion in the log file.