Fix buffer overflow in compressed & array texture hash calculation
The code that calculated the texture hashes had several incorrect
assumptions. In particular, it used the size of the texture in texels as
the buffer size, failed to account for compressed textures and treated
arrays of textures as a single large buffer instead of multiple smaller
buffers.

The result is that the texture CRC would almost always be based on an
incorrect amount of data. In the case where it was too short this was
not fatal - it would not be complete, but it's likely that it would be
good enough. In the case where it was too large things could be worse.
On the one hand, there could be nothing mapped after the buffer, which
would result in a data storage exception, which we would catch and
return 0 for the CRC. However, if there did happen to be something
mapped after the buffer we would treat it as part of the texture
resulting in an unpredictable and likely inconsistent CRC.

Problem is, there are now a number of games using the texture CRCs for
filtering, so fixing this would very likely break those fixes. This
patch makes a compromise - it checks the bad length against the actual
length of the first item in the texture array. If it is equal or
shorter, it will use the old length calculation and only the first item
in the array so that the CRC matches.

On the other hand, if it overflows it assumes that the CRC was going to
be broken/inconsistent/zero anyway and calculates it using the correct
size of the buffer of the first mip-map for all items in the texture
array.

I anticipate that most texture CRCs we are using in practice will fall
into the former category, so hopefully the pain of changing CRCs will be
minimal.