Fix D3D11CalcSubresource with mipmaps=0
I ran a quick test to see what D3D11CalcSubresource would do if mip-maps
== 0 (as when creating a texture and asking DX to create the mip-maps
automatically) but multiple textures are present in the array, and sure
enough it always returns 0 in that case, regardless of which item was
asked for.

Fix this by ensuring that mip-maps is always at least 1 when using this
utility function.