Small tweak to refcount fix to call original.
Slightly better to call back to the original, even though we don't use the result.  This makes it match the goal of always calling the original routine.

The reason to make this tweak is that other wrappers may wrap this call too, and calling the original keeps the code path consistent for them too, as well as bumping the refcount.

Remove some old documentation, and I think we are now sure that returning the wrapped device is what we want.