Frame Analysis: Dump strides and offsets for vertex and index buffers
The full vertex + index buffers are not that useful without knowing
what data in the buffers are relevant. There's a few pieces we will
ultimately need to piece this all together, of which this is the lowest
hanging fruit - the offsets and strides of the vertex and index buffers.

The vertex buffers dumps are still not ideal - the data is not
interpreted using the correct formats, but we can't determine that
unless we save off information from when the input assembler layout was
created.

We should also check the parameters passed to the draw calls for the
complete picture.