Add resolve_msaa feature to resource copying (not recommended for use yet)
If we are copying a resource from a shader working with anti-aliased
textures into one working with non-AA textures we may need to resolve
the anti-aliasing first (particularly when the destination is only
a shader model 4 shader).

This patch adds a new resolve_msaa keyword which calls out to
ResolveSubresource to handle some cases of these cases. Unfortunately,
not all cards can resolve all possible formats, and I don't have any
compatibility matrix to see if this how serious this is. All I know is
my card cannot resolve the formats typically used by depth buffers.

In the future we might be able to call a custom shader to resolve the
textures in cases where the card cannot do so directly, but until then
it's probably not a good idea to encourage people to use this feature
since we don't know what sort of compatibility issues people will hit in
the field.