Far Cry 4: Fix HBAO+ artefacts
Certain surfaces displayed some artifacts in 3D while using HBAO+. This
particularly affected large mostly flat surfaces that had a lot of
detail in their normal map - things like walls, large rocky surfaces and
the like. Grass and vegetation already looks perfect with HBAO+ (and is
the main reason to pick it over other AO algorithms).

Using frame analysis, I identified two shaders that generated artefacts
between the eyes, though this was more tricky than usual.

The first was difficult to spot as the artefacts were output in the W
channel, which is displayed as alpha in most image editors - something
to look out for, but that became apparent examining the inputs of the
second that something was already broken before that point.

The second was difficult to spot, as it outputs Texture3D targets which
frame analysis (and DirectTK) cannot currently dump. I inferred that
these shaders were to blame as the input to the following shader was
clearly broken, and no dumped render targets from the start of the HBAO
pass matched it - so it made sense that it came from the shaders that
failed to dump their render targets and were immediately before it was
seen on an input.

The artefacts from both of these shaders can be fixed by removing the
the view-space stereo formula from each 3D point generated from the W
buffer. To do this accurately and independently of the FOV, the
projection or inverse projection matrix must be copied into the HBAO+
shaders with 3DMigoto as they are not normally available there.

Fixing HBAO+ without access to the projection matrix is certainly
possible, but it would just be an approximation if the FOV can vary -
but still a lot better than leaving it unfixed.