Force return an E_INVALIDARG for any CreateDevice other than for D3D_FEATURE_LEVEL_11_0.
This seems really quite heavy handed, but has tested out to be successful in 13 games so far.

The idea is that we need to only wrap objects and create stereo textures and so on for DX11.0 devices only, and not the others that come our way during game launch.

Not sure why returning an error for bad requests like DX9.2 doesn't make the games mad, but so far they are not failing.

This fixes the hang at launch on Win10 for FarCry4.
It also makes the 1.2 branch work for Watch_Dogs, which has never worked before.

This is a little bold as a fix, but unless we find problems with it, it's off to a good start.

We can always make it an optional parameter in the d3dx.ini file if we run into problems.