Resource Copying: Handle extra constraints copying buffers to constant buffers
If copying a buffer into a constant buffer slot, there are some
additional restrictions. The buffer must be a multiple of 16 bytes, and
cannot contain more than 4096 x 4 component x 4 byte entries.

Further, unless we are using Windows 8, there is no way to set an offset
when binding the constant buffers to the pipeline, but we may need the
offset in order to find the right vertices in a vertex buffer, etc.

For both of these situations, use CopySubResourceRegion to copy the
an appropriate portion of the original buffer into the constant buffer,
taking the offset and size constraints into account.