Far Cry 4: The Ultimate Auto UI Depth
With the new ability to copy a vertex buffer into a constant buffer, we
can look up the position of other vertices in the hud shader. This gives
us the ability to find the position on the screen of each the current
HUD icon from each vertex, and use it to adjust the depth of the icon
consistently.

- Apply an adjustment to various HUD icons that point to items in the
  world based on the center of the icon.

- Blacklist icons in the bottom left of the screen where the minimap is

- Detect if the depth buffer is active or not and disable all UI
  adjustments when it is not (simplifies things)

This game is going to have the best HUD ever in 3D Vision!