Add support to use vertex buffers as structured buffers in SRV slots
Using a vertex buffer in a constant buffer slot works, but makes it
difficult to address each vertex in the shaders. If we instead inject it
as a structured buffer in a shader resource slot, it should be possible
to address directly.

This patch detects this situation, and sets the necessary flags and
creates the necessary SRV description to use vertex buffers as
structured buffers. Similar to constraints when using a buffer as a
constant buffer, it seems that structured buffers need to be a multiple
of their stride. The code takes care of setting an appropriate buffer
size and uses the same region copy code as constant buffers to ensure
the starting element is aligned at the start of the buffer.

It also has an additional constraint that the source region used in the
copy operation is also a multiple of the destination stride... take care
of that as well.

This does not address other cases of copying a buffer from one binding
type to another, which may still fail to create a view without a
description.