HLSL decompiler: Fix heap corruption when parsing hull, domain and geometry shaders
When the decompiler parses the hs_5_0, gs_5_0 or ds_5_0 lines from these
type of shaders it did not recognise them and treated them as an
instruction with opcode 0 (add). This will then cause heap corruption
later in applySwizzle() when it attempts to fails to find a "." in an
empty string and writes a 0 to a random address.