BinaryDecompiler: Replace ASSERTs we hit in practice with exceptions
asserts are compiled out in release code, but we hit these two in
practice in shaders with unsupported instructions. This can result in
UpdateOperandReferences using the uninitialised ui32NumOperands as the
length of the psInst->asOperands array, potentially crashing, or perhaps
just using garbage data in further logic.

Replace the assertions with a log message and raise exception to make
sure we bail out before any damage can be done.

This adds the repository root to the include paths for the binary
decompiler project so it can include log.h.