Far Cry 4: Use vertex buffer as a structured buffer instead of a constant buffer
Now that we can copy vertex buffers into a shader as a structured
buffer, we can drastically simplify the code to look up other vertex
positions - we no longer need to manually calculate the offset and
components of the xy position, we can simply look up the vertex buffer
directly by id and access the position in the struct.

The previous code worked fine, but this should be easier for others to
replicate.