Disable geometry + domain hunting key bindings by default
3DMigoto still processes input while alt+tabbed out of a game, and the
keys I chose for these are pretty common to press while e.g. editing a
shader or performing other tasks. It's possible to avoid that by
toggling hunting off, but it doesn't seem worthwhile yet - therefore,
disable the bindings by default once again.

Ultimately it would be better if we can find a way to detect if the
game's window is foregrounded and disable all the keys otherwise.