Resource Copying: Refactor CreateCompatibleView with templates
The DirectX APIs are somewhat annoying that they define almost but not
quite identical interfaces for different types of resources that we have
to deal with. This is of course bad because it leads to duplicated code
which leads to copy + paste bugs and a lack of copy + paste bug fixes.

C++ function templates and specialisations are one solution to this...
They may not be the prettiest things but at least they mean we can use a
common code path for the common code.

This refactors CreateCompatibleView to use common code between the four
view types, and support filling out the view description when coercing a
buffer to a structured buffer for resource views, render targets and
unordered access views (depth/stencil targets don't support buffers).