Add ability to inject index buffers to shaders
Similar to injecting vertex buffers, we might also want to inject an
index buffer into a shader. This is a simpler case as we don't need to
modify the existing buffer at all since it is a fixed format buffer, and
we can mostly reuse existing code, with some refactoring.