Resource copying: Add support stereo, mono and stereo2mono keywords
This adds two new keywords that can force the surface creation mode when
creating a resource during a resource copy.

It also adds a new "stereo2mono" keyword that will take a stereo 2D
texture resource and perform a reverse stereo blit to convert it into a
double width mono resource.

This can be used to give a shader access to the perspective of the other
eye, which should open up some interesting possibilities, particularly
for the auto-depth adjustment shaders.

It seems that the reverse stereo blit requires the destination resource
to be stereo - if the destination is mono it will behave as though it is
not enabled. However this is a bit of a problem since the destination,
while stereo, only contains the copy in the right eye. To solve this, we
use two steps - we first perform the reverse stereo blit to an
intermediate texture that we have forced to stereo, then copy that to
the final destination which we force to mono.

In the event that the reverse stereo blit fails to enable (e.g. if
stereo is disabled in the control panel), we call CopySubresourceRegion
twice to copy the original 2D resource into both sides of the 2x width
destination.

Note that there are some situations we found where reverse stereo blit
failed in frame analysis that we need to check - in particular, copying
from an MSAA texture, or a depth buffer. Need to confirm if they are
problems here, or if that was to do with the staging texture we use for
frame analysis.