Add command lists to TextureOverride sections
This gives us the ability to set ini params and use arbitrary resource
copying from a texture override section, just as they can be from
ShaderOverride and Present sections.

Because it would be a performance nightmare to check texture overrides
for every texture slot of every shader on every draw call, the command
list must be invoked from a command in a ShaderOverride section, e.g.:

[ShaderOverride1]
Hash = ...
checktextureoverride = ps-t0
checktextureoverride = ps-t1

[TextureOverride1]
Hash = ...
x = 1

I have a few obscure needs for this, but it will become most useful
later as part of scene detection.

This reuses the TextureOverride lookup code from the ini param texture
filtering feature and refactors the command list structures a little so
the same command can be added to both pre and post lists (so a post line
in a TextureOverride section will work without having to remember to do
a 'post checktextureoverride' in the ShaderOverride as well).