Far Cry 4: Fix DLC exclamation marks on main menu
The exclamation marks next to the DLC items on the main menu use the
same vertex shader, pixel shader and texture as the highlighted cross on
the camera HUD, so there's no direct way to determine what it is being
used for.

Use the new ability to run command lists from TextureOverride sections
for scene detection within the frame - when the camera cross texture is
active set a flag which we can then check from the HUD shader to see if
the camera HUD has already been drawn this frame.