Far Cry 4: Note something weird with the camera HUD textures
There's something odd going on with some of the textures in this game,
which are changing hashes randomly - sometimes just opening and closing
the camera will be enough to change a texture hash, but it seems much
more likely after an alt+tab.

The textures are binary identical when dumped out using frame analysis
as .dds files, and debug shows their desc structures do not change, but
it seems to indicate that their initial data does.

More investigation is required - it might be that the game is creating
them with garbage and updating them later, or we could have a bug
somewhere. These textures are BC3 compressed, but the size calculation
we use for the hash of 16K matches up with the dumped .dds files.