Track resource updates that may contaminate the hash in ShaderUsage.txt
Far Cry 4 seems to like copying textures around on the GPU with
CopySubresourceRegion for some reason even if they don't change. I can
only assume that it must be doing some form of vram optimisation or
streaming, but the result is that a resource may be altered at any time
and the hash will no longer be accurate for the texture they contain.

We can track this sort of behaviour in the ShaderUsage.txt. Any time a
resource is updated with UpdateSubresource(), CopyResource() or
CopySubresourceRegion() we mark the hash as potentially contaminated if
it originally was generated using an initial data. If it was copied from
another resource we record the source resource hash and if it was a
region copy we note whether it copied the full resource or only part of
it, and the minimum dimensions ever seen copied to a resource.

The result is the ShaderUsage.txt will now give us a reasonable
indication as to whether we can rely on a texture hash or not.