Dump ShaderUsage.txt on frame analysis
Since ShaderUsage is now tracking resource copies it can be useful to
cross-reference the dumped resources with the ShaderUsage, so write out
a copy of it in the frame analysis directory. Write it after the frame
analysis has completed to ensure it contains all information from the
frame being analysed.

This also merges the StartFrameAnalysis and AnalyseFrame functions,
since they are both called from the same Present call these days.