Ignore updates/copies that only replace a mip-map
Since the hash code explicitly ignores mip-maps, we don't need to worry
about updates to them changing the texture hash, so bail out early when
checking for hash contamination.

This also refactors the code so it records whether a mip-map copy was
partial or full accurately, to confirm my understanding and so we can
collect info on mip-map updates if we ever needed to in the future -
but, as above this is disabled for now.

The subresource updates I'm seeing do raise a question as to how we
should handle updates to a texture with multiple array elements. For
simple resource copies we could use strategies to update the hash during
the copy/update/map, but this becomes problematic if we needed to update
a hash of a multi-element texture when only a single subresource was
being updated. We could consider a new strategy for hashing textures to
solve this - we could hash each subresource individually and combine all
the subhashes with an xor - that way we could replace any individual
subhash quite easily.

This is largely academic for now - FC4 does use a number of
multi-element resources, but so far none of them are showing as being
updated from the CPU. Still worth thinking about though...